Funding & Collateral calc (#21)
* feat: twap oracle

* clean: remove interface and replace with @uniswap

* fix: controller unit tests

* unit-test: init pool for oracle testing

* unit-test: TWAP

* chore: change functions order

* feat: get mark and index price

* fix

* fix

* fix: mark price calculation

* clean test

* fix: fetching TWAP

* chore: rename function

* chore: rename squeeth to wrapped squeeth

* chore: re-order import statements

* feat: funding calc

* feat: collateral calc

* fix: collat calc

* chore: remove division

* chore: add uniswap mock

* chore: fic typo

* chore: add vault function in VaultLib

* update open vault

* fix collat check

* fix collat check

* divide minted amount by normalizedFactor

* update funding equation

* chore

* chore

* chore: rename var

* fix typo


* add comment


* update oracle mock

* rename vars'

* more code

* update funding equation

* fix normalizationFactor

* fix opening vault with no collateral

* fix funding

* fix calc

* fix _isProperlyCollateralized

* fix test

* change uint128 to uint256

* feat: liquidate()

* fix compile

* FIxes to Contoller Tests (#31)

* first test WIP

* fixed tests, oracle values, and mint/collat params

* removing incorrect merge


* linting

* Fix the system can't be initiated because minting will not work while pool is empty (#30)

* temp fix to make system work with period > 1

* fix: don't quote twap when pool is empty

* fix: don't quote twap when pool is empty

* fix remaining tests

* lint: contracts

+++ packages/hardhat/contracts/periphery/ShortHelper.sol
@@ -70,7 +70,7 @@ contract ShortHelper {
      */
     function closeShort(
         uint256 _vaultId,
-        uint128 _burnSqueethAmount,
+        uint256 _burnSqueethAmount,
         uint128 _withdrawAmount,
         ISwapRouter.ExactOutputSingleParams memory _exactOutputParams
     ) external payable {
